+++ src/bms/BMS1.java
@@ -19,6 +19,41 @@ public class BMS1 extends BMSSecondary {
      */
 
     /**
+     * Default size of a blade pack.
+     */
+    private static final int DEFAULT_SIZE = 100;
+
+    /**
+     * Length of the sub array.
+     */
+    private static final int LENGTH_DATA = 5;
+
+    /**
+     * Index of voltage in the returned blade array.
+     */
+    private static final int VOLTAGE_INDEX = 0;
+
+    /**
+     * Index of current in the returned blade array.
+     */
+    private static final int CURRENT_INDEX = 1;
+
+    /**
+     * Index of temprature in the returned blade array.
+     */
+    private static final int TEMP_INDEX = 2;
+
+    /**
+     * Index of soc in the returned blade array.
+     */
+    private static final int SOC_INDEX = 3;
+
+    /**
+     * Index of soh in the returned blade array.
+     */
+    private static final int SOH_INDEX = 4;
+
+    /**
      * Representation of all blade battery slots.
      */
     private double[][] rep;
@@ -47,7 +82,7 @@ public class BMS1 extends BMSSecondary {
      * No-argument constructor.
      */
     public BMS1() {
-        this.createNewRep(100);
+        this.createNewRep(DEFAULT_SIZE);
     }
 
     /**
@@ -78,7 +113,7 @@ public class BMS1 extends BMSSecondary {
 
     @Override
     public final void clear() {
-        this.createNewRep(100);
+        this.createNewRep(DEFAULT_SIZE);
     }
 
     @Override
@@ -92,11 +127,72 @@ public class BMS1 extends BMSSecondary {
         this.rep = localSource.rep;
         this.activeSize = localSource.activeSize;
 
-        localSource.createNewRep(100);
+        localSource.createNewRep(DEFAULT_SIZE);
     }
+
     /*
      * Kernel methods ---------------------------------------------------------
      */
 
-    // TODO: insertBlade, updateData, removeBlade, size, bladeAt
+    @Override
+    public final void insertBlade(int index, double voltage, double current,
+            double temperature, int soc, double soh) {
+        assert index >= 0 : "Violation of: index >= 0";
+        assert index < this.rep.length : "Violation of: index < max capacity";
+        assert this.rep[index] == null : "Violation of: slot is currently empty";
+
+        double[] newBlade = { voltage, current, temperature, soc, soh };
+        this.rep[index] = newBlade;
+
+        this.activeSize++;
+    }
+
+    @Override
+    public final void updateData(int index, double newVoltage,
+            double newCurrent, double newTemperature, int newSoc,
+            double newSoh) {
+        assert index >= 0 : "Violation of: index >= 0";
+        assert index < this.rep.length : "Violation of: index < max capacity";
+        assert this.rep[index] != null : "Violation of: slot contains an active battery";
+
+        this.rep[index][VOLTAGE_INDEX] = newVoltage;
+        this.rep[index][CURRENT_INDEX] = newCurrent;
+        this.rep[index][TEMP_INDEX] = newTemperature;
+        this.rep[index][SOC_INDEX] = newSoc;
+        this.rep[index][SOH_INDEX] = newSoh;
+    }
+
+    @Override
+    public final double[] removeBlade(int index) {
+        assert index >= 0 : "Violation of: index >= 0";
+        assert index < this.rep.length : "Violation of: index < max capacity";
+        assert this.rep[index] != null : "Violation of: slot contains an active battery";
+
+        double[] removedData = this.rep[index];
+
+        this.rep[index] = null;
+        this.activeSize--;
+
+        return removedData;
+    }
+
+    @Override
+    public final int size() {
+        return this.activeSize;
+    }
+
+    @Override
+    public final double[] bladeAt(int index) {
+        assert index >= 0 : "Violation of: index >= 0";
+        assert index < this.rep.length : "Violation of: index < max capacity";
+        assert this.rep[index] != null : "Violation of: slot contains an active battery";
+
+        double[] copiedBlade = new double[LENGTH_DATA];
+
+        for (int i = 0; i < LENGTH_DATA; i++) {
+            copiedBlade[i] = this.rep[index][i];
+        }
+
+        return copiedBlade;
+    }
 }